+++ test/core/package_unique.sv
@@ -1,0 +1,9 @@
+package P;
+    typedef logic T;
+endpackage
+
+module top;
+    P::T P_T;
+    assign P_T = 0;
+    initial $display("%b", P_T);
+endmodule