+++ test/core/package_unique.v
@@ -1,0 +1,5 @@
+module top;
+    wire P_T;
+    assign P_T = 0;
+    initial $display("%b", P_T);
+endmodule